+++ SpicesUpdate@claudiux/files/SpicesUpdate@claudiux/4.2/constants.js
@@ -24,7 +24,7 @@ function DEBUG() {
 const DOWNLOAD_TIME = 10;
 
 let sort = "date";
-if (versionCompare( GLib.getenv('CINNAMON_VERSION') ,"4.3.0" ) >= 0 ) {
+if (versionCompare( GLib.getenv('CINNAMON_VERSION') ,"4.2.4" ) >= 0 ) {
   sort = "update";
 }