+++ SpicesUpdate@claudiux/files/SpicesUpdate@claudiux/cs/4.2/modules/cs_applets.py
@@ -1,7 +1,7 @@
 #!/usr/bin/python3
 
 import sys
-from ExtensionCore import ManageSpicesPage, DownloadSpicesPage
+from SUExtensionCore import ManageSpicesPage, DownloadSpicesPage
 from SettingsWidgets import SidePage, SettingsStack
 from SUSpices import SU_Spice_Harvester
 from gi.repository import GLib, Gtk, Gdk